+++ e2e/fixtures/extensions/extension-add-config/simple-config/simple-config.manifest.js
@@ -1,0 +1,15 @@
+const ComponentFactoryExt = require('bit-bin').ComponentFactoryExt;
+
+module.exports = {
+  name: 'simple config',
+  dependencies: [ComponentFactoryExt],
+  provider: async ([component]) => {
+    console.log('simple config runs');
+    component.registerAddConfig('simple-config', config => {
+      console.log('config registration hook is running');
+      return {
+        'my-custom-key': 'my-custom-val'
+      };
+    });
+  }
+};