+++ e2e/flows/out-of-sync-componets.e2e.3.ts
@@ -1,6 +1,6 @@
 import chai, { expect } from 'chai';
 import Helper from '../../src/e2e-helper/e2e-helper';
-import { statusWorkspaceIsCleanMsg, importPendingMsg } from '../../src/cli/commands/public-cmds/status-cmd';
+import { importPendingMsg } from '../../src/cli/commands/public-cmds/status-cmd';
 import { MissingBitMapComponent } from '../../src/consumer/bit-map/exceptions';
 import ComponentsPendingImport from '../../src/consumer/component-ops/exceptions/components-pending-import';
 
@@ -89,13 +89,12 @@ describe('components that are not synced between the scope and the consumer', fu
       });
     });
     describe('bit status', () => {
-      let output;
       before(() => {
         helper.scopeHelper.getClonedLocalScope(scopeOutOfSync);
-        output = helper.command.status();
+        helper.command.status();
       });
       it('should sync .bitmap according to the scope', () => {
-        expect(output).to.have.string(statusWorkspaceIsCleanMsg);
+        helper.command.expectStatusToBeClean();
         const bitMap = helper.bitMap.read();
         const newId = `${helper.scopes.remote}/bar/foo@0.0.1`;
         expect(bitMap).to.have.property(newId);
@@ -122,13 +121,12 @@ describe('components that are not synced between the scope and the consumer', fu
       });
     });
     describe('bit status', () => {
-      let output;
       before(() => {
         helper.scopeHelper.getClonedLocalScope(scopeOutOfSync);
-        output = helper.command.status();
+        helper.command.status();
       });
       it('should sync .bitmap according to the scope', () => {
-        expect(output).to.have.string(statusWorkspaceIsCleanMsg);
+        helper.command.expectStatusToBeClean();
         const bitMap = helper.bitMap.read();
         const newId = `${helper.scopes.remote}/bar/foo@0.0.1`;
         expect(bitMap).to.have.property(newId);
@@ -375,12 +373,11 @@ describe('components that are not synced between the scope and the consumer', fu
         scopeOutOfSync = helper.scopeHelper.cloneLocalScope();
       });
       describe('bit status', () => {
-        let output;
         before(() => {
-          output = helper.command.status();
+          helper.command.status();
         });
         it('should sync .bitmap according to the latest version of the scope', () => {
-          expect(output).to.have.string(statusWorkspaceIsCleanMsg);
+          helper.command.expectStatusToBeClean();
           const bitMap = helper.bitMap.read();
           const newId = `${helper.scopes.remote}/bar/foo@0.0.1`;
           expect(bitMap).to.have.property(newId);