+++ e2e/functionalities/binary-files.e2e.2.ts
@@ -3,7 +3,6 @@ import fs from 'fs-extra';
 import glob from 'glob';
 import chai, { expect } from 'chai';
 import Helper from '../../src/e2e-helper/e2e-helper';
-import { statusWorkspaceIsCleanMsg } from '../../src/cli/commands/public-cmds/status-cmd';
 import NpmCiRegistry, { supportNpmCiRegistryTesting } from '../npm-ci-registry';
 import { AUTO_GENERATED_STAMP } from '../../src/constants';
 
@@ -135,8 +134,7 @@ describe('binary files', function() {
       expect(symlinkValue).to.be.a.path();
     });
     it('bit-status should not show the component as modified', () => {
-      const status = helper.command.status();
-      expect(status).to.have.string(statusWorkspaceIsCleanMsg);
+      helper.command.expectStatusToBeClean();
     });
   });
   describe('import a PNG file as a dependency with custom-resolve-modules', () => {
@@ -174,8 +172,7 @@ describe('binary files', function() {
       expect(symlinkValue).to.be.a.path();
     });
     it('bit-status should not show the component as modified', () => {
-      const status = helper.command.status();
-      expect(status).to.have.string(statusWorkspaceIsCleanMsg);
+      helper.command.expectStatusToBeClean();
     });
     (supportNpmCiRegistryTesting ? describe : describe.skip)('when dependencies are saved as packages', () => {
       let barFooPath;
@@ -240,8 +237,7 @@ describe('binary files', function() {
       expect(symlinkValue).to.be.a.path();
     });
     it('bit-status should not show the component as modified', () => {
-      const status = helper.command.status();
-      expect(status).to.have.string(statusWorkspaceIsCleanMsg);
+      helper.command.expectStatusToBeClean();
     });
     (supportNpmCiRegistryTesting ? describe : describe.skip)('when dependencies are saved as packages', () => {
       let barFooPath;
@@ -327,8 +323,7 @@ describe('binary files', function() {
       expect(symlinkValue).to.be.a.path();
     });
     it('bit-status should not show the component as modified', () => {
-      const status = helper.command.status();
-      expect(status).to.have.string(statusWorkspaceIsCleanMsg);
+      helper.command.expectStatusToBeClean();
     });
     (supportNpmCiRegistryTesting ? describe : describe.skip)('when dependencies are saved as packages', () => {
       let barFooPath;
@@ -426,8 +421,7 @@ describe('binary files', function() {
       expect(symlinkValue).to.be.a.path();
     });
     it('bit-status should not show the component as modified', () => {
-      const status = helper.command.status();
-      expect(status).to.have.string(statusWorkspaceIsCleanMsg);
+      helper.command.expectStatusToBeClean();
     });
     (supportNpmCiRegistryTesting ? describe : describe.skip)('when dependencies are saved as packages', () => {
       let barFooPath;